Web WidthProvider
Consolidated the web and native width providers. This was necessary
since react native is only including the "index.js" and will never see
any of the ".web.js" extensions within the source directory. Inlined the
babel configuration so that react native does not attempt to apply it to
the minified source.